The lifecycle of an advice works as expected

+++ Workspace/org.collin.moodle/src/org/collin/moodle/advice/AdviceMap.java
@@ -60,7 +60,7 @@ public class AdviceMap implements IAdviceMap {
 	}
 
 	@Override
-	public long getAdviceId() {
+	public int getAdviceId() {
 		return adviceId;
 	}
