+++ Workspace/org.collin.moodle/src/org/collin/moodle/advice/IAdviceMap.java
@@ -28,7 +28,7 @@ public interface IAdviceMap {
 
 	Interactions getInteraction();
 
-	long getAdviceId();
+	int getAdviceId();
 
 	Notifications getNotification();
